requirements.txt
fastapi==0.115.0
python-dotenv==1.0.1
pydantic==2.9.2
langdetect==1.0.9
httpx==0.27.2
groq==0.11.0
gradio_client==1.3.0